@@ -1,4 +1,5 @@
 using pueblo_golf_tournament_api.Base;
+using pueblo_golf_tournament_api.Utilities.Enums;
 
 namespace pueblo_golf_tournament_api.Dto
 {
@@ -7,7 +8,7 @@ namespace pueblo_golf_tournament_api.Dto
         public string? Username { get; set; }
         public string? Password { get; set; }
         public long PersonId { get; set; }
-
+        public AccountTypeEnum AccountType { get; set; }
         public virtual PersonDto? Person { get; set; }
     }
     public class AccountRawDto
