@@ -4,6 +4,11 @@ namespace pueblo_golf_tournament_api.Dto.Outgoing
 {
     public class RegisteredAccountDto : BaseResponse
     {
-        public PersonDto? Data { get; set; }
+        public RegisteredAccountDataDto? Data { get; set; }
+    }
+
+    public class RegisteredAccountDataDto {
+        public PersonDto? Person { get; set; }
+        public PlayerDto? Player { get; set; }
     }
 }
