@@ -9,12 +9,11 @@ namespace pueblo_golf_tournament_api.Dto
         public string? WorldHandicapSystemId { get; set; }
         public double Handicap { get; set; }
         public double DivisionId { get; set; }
-        public double HomeClubId { get; set; }
+        public string? HomeClub { get; set; }
         public double TeamId { get; set; }
         public double PersonId { get; set; }
         public virtual PersonDto? Person { get; set; }
         public virtual DivisionDto? Division { get; set; }
-        public virtual HomeClubDto? HomeClub { get; set; }
         public virtual TeamDto? Team { get; set; }
     }
     
@@ -22,6 +21,7 @@ namespace pueblo_golf_tournament_api.Dto
     {
         public string? WorldHandicapSystemId { get; set; }
         public double Handicap { get; set; }
+        public string? HomeClub { get; set; }
     }
 
 }
