@@ -9,11 +9,10 @@ namespace pueblo_golf_tournament_api.Entities
         public string? WorldHandicapSystemId { get; set; }
         public double Handicap { get; set; }
         public double DivisionId { get; set; }
-        public double HomeClubId { get; set; }
+        public string HomeClub { get; set; }
         public double TeamId { get; set; }
         public double PersonId { get; set; }
         public virtual Person? Person { get; set; }
-        public virtual HomeClub? HomeClub { get; set; }
         public virtual Team? Team { get; set; }
     }
 }
