@@ -6,6 +6,7 @@ using pueblo_golf_tournament_api.Dto.Outgoing;
 using pueblo_golf_tournament_api.Extensions;
 using pueblo_golf_tournament_api.Services.Accounts;
 using pueblo_golf_tournament_api.Services.Persons;
+using pueblo_golf_tournament_api.Services.Players;
 
 namespace pueblo_golf_tournament_api.Modules.Authentications
 {
@@ -13,13 +14,15 @@ namespace pueblo_golf_tournament_api.Modules.Authentications
     {
         private readonly IAccountService _accountService;
         private readonly IPersonService _personService;
+        private readonly IPlayerService _playerService;
         private readonly IMapper _mapper;
 
-        public AuthModule(IMapper mapper, IAccountService accountService, IPersonService personService)
+        public AuthModule(IMapper mapper, IAccountService accountService, IPersonService personService, IPlayerService playerService)
         {
             _mapper = mapper;
             _accountService = accountService;
             _personService = personService;
+            _playerService = playerService;
         }
 
         public async Task<AuthenticationLoginDto> Login(AuthenticateLoginDto payload)
@@ -53,10 +56,12 @@ namespace pueblo_golf_tournament_api.Modules.Authentications
             }
 
             var person = await _personService.GetAsync(person => person.Id.Equals(account.PersonId));
+            var player = await _playerService.GetAsync(player => player.PersonId.Equals(account.PersonId));
 
             response.Data = new AuthenticatedUserData
             {
                 PersonalDetails = _mapper.Map<PersonDto>(person),
+                PlayerDetails = _mapper.Map<PlayerDto>(player),
                 Username = account.Username
             };
             response.Authenticated = true;
