@@ -7,13 +7,23 @@ part of 'account.dart';
 // **************************************************************************
 
 Account _$AccountFromJson(Map<String, dynamic> json) => Account(
-      id: (json['id'] as num).toDouble(),
       username: json['username'] as String,
       password: json['password'] as String,
+      personId: (json['personId'] as num).toDouble(),
+      accountType: $enumDecode(_$AccountTypeEnumEnumMap, json['accountType']),
+      id: json['id'] as int,
     );
 
 Map<String, dynamic> _$AccountToJson(Account instance) => <String, dynamic>{
       'id': instance.id,
       'username': instance.username,
       'password': instance.password,
+      'personId': instance.personId,
+      'accountType': _$AccountTypeEnumEnumMap[instance.accountType]!,
     };
+
+const _$AccountTypeEnumEnumMap = {
+  AccountTypeEnum.SuperAdministrator: 'SuperAdministrator',
+  AccountTypeEnum.Administrator: 'Administrator',
+  AccountTypeEnum.User: 'User',
+};
