@@ -1,26 +1,10 @@
-import 'dart:math';
-
-import 'package:currency_text_input_formatter/currency_text_input_formatter.dart';
-
-import 'package:flutter/material.dart';
-import 'package:get/get.dart';
-import 'package:get/get_core/src/get_main.dart';
-
-import 'package:fl_chart/fl_chart.dart';
-import 'package:fl_chart/fl_chart.dart';
-import 'package:flutter/material.dart';
-import 'package:pueblo_golf_tournament_mobile/widgets/tournament-banner.dart';
+import 'package:flutter/widgets.dart';
 
 class MenuPage extends StatelessWidget {
-  MenuPage({Key? key}) : super(key: key);
-
   @override
   Widget build(BuildContext context) {
     return SingleChildScrollView(
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [],
-      ),
+      child: Column(children: []),
     );
   }
 }
